+++ packages/fabric/editor-core/src/editor/types/editor-props.ts
@@ -20,6 +20,7 @@ import { TextFormattingOptions } from '../plugins/text-formatting';
 import { CollabEditProvider } from '../plugins/collab-edit';
 import { MacroProvider } from '../plugins/macro/types';
 import { MediaOptions } from '../plugins/media';
+import { PluginConfig as TablesPluginConfig } from '../../plugins/table';
 
 export type EditorAppearance =
   | 'message'
@@ -51,6 +52,7 @@ export interface EditorProps {
   secondaryToolbarComponents?: ReactElement;
   addonToolbarComponents?: ReactElement;
 
+  allowBlockType?: { exclude?: Array<string> };
   allowTextFormatting?: boolean | TextFormattingOptions;
   allowMentions?: boolean;
   allowTasksAndDecisions?: boolean;
@@ -59,7 +61,7 @@ export interface EditorProps {
   allowCodeBlocks?: boolean;
   allowLists?: boolean;
   allowTextColor?: boolean;
-  allowTables?: boolean;
+  allowTables?: boolean | TablesPluginConfig;
   allowHelpDialog?: boolean;
   allowJiraIssue?: boolean;
   allowUnsupportedContent?: boolean;
@@ -68,7 +70,6 @@ export interface EditorProps {
   allowConfluenceInlineComment?: boolean;
   allowPlaceholderCursor?: boolean;
   allowDate?: boolean;
-  allowTableColumnResizing?: boolean;
 
   saveOnEnter?: boolean;
   shouldFocus?: boolean;