+++ packages/fabric/editor-json-transformer/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @atlaskit/editor-json-transformer
 
+## 2.5.2
+
 ## 2.5.1
 
 ## 2.5.0